fix(tethys): address PR #57 review feedback
- Switch get_impact / get_symbol_impact max_depth to Option<usize> for
  consistency with get_forward_reachable / get_backward_reachable
- Re-export the existing db::graph::DEFAULT_MAX_DEPTH constant so the
  default value (50) lives in one place; reuse it in both methods
- Clarify in doc comments that None means default-50, not unbounded
  traversal
- Tighten get_impact_max_depth_limits_transitive_traversal: assert
  exact direct/transitive counts on the deterministic call_graph fixture
  instead of a loose <= comparison
- Add a parallel get_symbol_impact_max_depth_limits_transitive_traversal
  test for symmetry
- Remove the stale "(not yet implemented)" doc on the --depth CLI arg
- Note in cli/callers.rs why None is passed to get_symbol_impact

Refs: rivets-w0qw (depth=0 semantics — tracked as a follow-up)

+++ src/cli/impact.rs
@@ -13,7 +13,7 @@ pub fn run(
     workspace: &Path,
     target: &str,
     is_symbol: bool,
-    depth: Option<u32>,
+    depth: Option<usize>,
     lsp: bool,
 ) -> Result<(), tethys::Error> {
     ensure_lsp_if_requested(lsp)?;
